refactor : Skill 클래스 위치 이동
