+++ src/main/java/univ/kgu/carely/domain/common/embeded/skill/Skill.java
@@ -1,4 +1,4 @@
-package univ.kgu.carely.domain.common.embeded;
+package univ.kgu.carely.domain.common.embeded.skill;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.Embeddable;
